chore(CarRepository): 클래스 삭제
