+++ src/main/java/model/car/Cars.java
@@ -3,6 +3,7 @@ package model.car;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class Cars {
     private final List<Car> cars = new ArrayList<>();
@@ -13,5 +14,20 @@ public class Cars {
 
     public void addCar(Car car) {
         cars.add(car);
+    }
+
+    public List<String> getWinners() {
+        int maxPosition = getMaxPosition();
+        return cars.stream()
+                .filter(car -> car.isWinner(maxPosition))
+                .map(Car::getName)
+                .collect(Collectors.toList());
+    }
+
+    private int getMaxPosition() {
+        List<Integer> positions = cars.stream()
+                .map(Car::getPosition)
+                .collect(Collectors.toList());
+        return Collections.max(positions);
     }
 }
